+++ Source/ACE/Factories/GenericObjectFactory.cs
@@ -18,12 +18,12 @@ namespace ACE.Factories
                 var oType = (ItemType)aceO.GetIntProperty(PropertyInt.ItemType);
                 var oDescFlag = (ObjectDescriptionFlag)aceO.AceObjectDescriptionFlags;
 
-                if (aceO.GetBoolProperty(PropertyBool.GeneratorStatus) ?? false)  // Generator
-                {
-                    aceO.SetPosition(PositionType.Location, aceO.GetPosition(PositionType.Location).InFrontOf(-2.0));
-                    aceO.GetPosition(PositionType.Location).PositionZ = aceO.GetPosition(PositionType.Location).PositionZ - 0.5f;
+                if (aceO.GeneratorStatus ?? false)  // Generator
+                {                    
+                    aceO.Location = aceO.Location.InFrontOf(-2.0);
+                    aceO.Location.PositionZ = aceO.Location.PositionZ - 0.5f;
                     results.Add(new Generator(new ObjectGuid(aceO.AceObjectId), aceO));
-                    aceO.SetBoolProperty(PropertyBool.GeneratorEnteredWorld, true);
+                    aceO.GeneratorEnteredWorld = true;
                     var objectList = GeneratorFactory.CreateWorldObjectsFromGenerator(aceO) ?? new List<WorldObject>();
                     objectList.ForEach(o => results.Add(o));
                     continue;
@@ -51,7 +51,7 @@ namespace ACE.Factories
                     default:
                         // Use the DebugObject to assist in building proper objects for weenies
                         // FIXME(ddevec): Some objects have null location.  This freaks out the landblock... ignore them?
-                        if (aceO.GetPosition(PositionType.Location) != null)
+                        if (aceO.Location != null)
                         {
                             results.Add(new DebugObject(aceO));
                         }