+++ Source/ACE/Factories/MonsterFactory.cs
@@ -17,9 +17,11 @@ namespace ACE.Factories
         /// </summary>
         public static Monster SpawnMonster(AceObject aceO, Position position)
         {
-            aceO.SetPosition(PositionType.Location, position);
+            // aceO.Location = position;
             // aceO.PhysicsScript = (ushort)Network.Enum.PlayScript.Create;
             Monster newMonster = new Monster(aceO);
+            newMonster.Location = position;
+            newMonster.GeneratorId = aceO.GeneratorIID;
 
             // newMonster.PhysicsData.DefaultScript = aceO.PhysicsScript;
             // newMonster.DefaultScript = (uint)Network.Enum.PlayScript.Create;