+++ Source/ACE/Network/Handlers/CharacterHandler.cs
@@ -30,6 +30,8 @@ namespace ACE.Network.Handlers
         public static void CharacterEnterWorld(ClientMessage message, Session session)
         {
             ObjectGuid guid = message.Payload.ReadGuid();
+            guid.ChangeGuidType(GuidType.Player);
+
             string account = message.Payload.ReadString16L();
 
             if (account != session.Account)
@@ -134,7 +136,7 @@ namespace ACE.Network.Handlers
             ////    CharacterCreateEx(message, session, id);
             ////});
 
-            uint id = GuidManager.NewPlayerGuid();
+            uint id = GuidManager.NewPlayerGuid().Full;
 
             CharacterCreateEx(message, session, id);
         }
@@ -318,7 +320,7 @@ namespace ACE.Network.Handlers
                     // DatabaseManager.Shard.SaveCharacterOptions(character);
                     // DatabaseManager.Shard.InitCharacterPositions(character);
 
-                    var guid = new ObjectGuid(character.AceObjectId);
+                    var guid = new ObjectGuid(character.AceObjectId, GuidType.Player);
                     session.AccountCharacters.Add(new CachedCharacter(guid, (byte)session.AccountCharacters.Count, character.Name, 0));
 
                     SendCharacterCreateResponse(session, CharacterGenerationVerificationResponse.Ok, guid, character.Name);